Added files via upload
Info - I own a lot of Books and Graphic Novels, all of which I have never catalogued. So why not create a little piece of software to help me do this? Well this is the first installment of Cataloguer; a small piece of software which aids in cataloguing a collection of Books, Novels, Comics, Graphic Novels etc... and saves it all in a small '.txt' file called 'catalogue.txt'.

Usage - Just fill in all the fields from 'Name', all the way through to 'Issues Owned' then click save to save that single entry, then click 'Clear' and re-fill all the fields for the next Book, Comic etc...

This is just my first installment, as I will be upgrading it's features